+++ command-palette/settings.js
@@ -118,8 +118,8 @@ const SettingsModal = {
     for (const [commandKey, shortcut] of Object.entries(
       this._currentSettings.customShortcuts || {}
     )) {
-      // Only save if: shortcut is not a default shortcut
-      if (shortcut && defaultShortcuts[commandKey] !== shortcut) {
+      // Only save if default shortcut is changed/removed
+      if (shortcut !== undefined && defaultShortcuts[commandKey] !== shortcut) {
         filteredCustomShortcuts[commandKey] = shortcut;
       }
     }
@@ -190,7 +190,10 @@ const SettingsModal = {
 
       // Remove old shortcuts
       for (const [commandKey] of Object.entries(oldShortcuts)) {
-        if (!newShortcuts[commandKey] && oldShortcuts[commandKey] !== defautlShortcut[commandKey]) {
+        if (
+          (!newShortcuts[commandKey] && oldShortcuts[commandKey] !== defautlShortcut[commandKey]) ||
+          newShortcuts[commandKey] === ""
+        ) {
           this._mainModule.removeHotkey(commandKey);
         }
       }
@@ -447,7 +450,8 @@ const SettingsModal = {
     if (key === "Backspace" || key === "Delete") {
       targetInput.value = "";
       if (commandKey) {
-        delete this._currentSettings.customShortcuts[commandKey];
+        // Set to empty string to indicate "unbound" instead of deleting
+        this._currentSettings.customShortcuts[commandKey] = "";
       }
       clearConflict();
       window.removeEventListener("keydown", this._boundHandleShortcutKeyDown, true);
@@ -486,7 +490,7 @@ const SettingsModal = {
         `Shortcut conflict detected for "${commandKey}" with shortcut "${shortcutString}":`,
         conflictCheck.conflicts
       );
-      delete this._currentSettings.customShortcuts[commandKey];
+      this._currentSettings.customShortcuts[commandKey] = shortcutString;
     } else {
       clearConflict();
       this._currentSettings.customShortcuts[commandKey] = shortcutString;
@@ -1111,13 +1115,13 @@ const SettingsModal = {
             </div>
             <div id="settings-tab-content" class="cmd-settings-tab-content" hidden>
               <!-- Content will be populated by _populateSettingsTab -->
-            </div>
+              </div>
             <div id="custom-commands-tab-content" class="cmd-settings-tab-content" hidden>
               <!-- Content will be populated by _populateCustomCommandsTab -->
-            </div>
+              </div>
             <div id="help-tab-content" class="cmd-settings-tab-content" hidden>
               <!-- Content will be populated by _populateHelpTab -->
-            </div>
+              </div>
           </div>
         </div>
       </div>