+++ command-palette/release-notes.md
@@ -1,7 +1,11 @@
-# New Features
-
 # Fixes
 
-# Others
+- Removing default shortcuts.
+- Event Listener Race Condition (Shortcut will not be triggered when input is in focus)
+
+# Breaking Changes
+- Command palette now allows saving shortcuts with conflict, It will give warning before saving.
 
 # Contributes
+
+@okilonet1 made his first contribution fixing shortcuts.